misc LLVM 3.1 warnings
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@9404 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ WebPortalConnection+Data.mm
@@ -1406,7 +1406,7 @@ const NSString* const GenerateMovieDicomImagesParamKey = @"dicomImageArray";
 			NSLog( @"****** WADO Server : study not found");
 		
 		if ([studies count] > 1)
-			NSLog( @"****** WADO Server : more than 1 study with same uid : %d", studies.count);
+			NSLog( @"****** WADO Server : more than 1 study with same uid : %d", (int) studies.count);
 		
         NSArray *allSeries = [NSArray array];
         
@@ -1513,7 +1513,7 @@ const NSString* const GenerateMovieDicomImagesParamKey = @"dicomImageArray";
 				NSLog( @"****** WADO Server : study not found");
 			
 			if ([studies count] > 1)
-				NSLog( @"****** WADO Server : more than 1 study with same uid : %d", studies.count);
+				NSLog( @"****** WADO Server : more than 1 study with same uid : %d", (int) studies.count);
 			
             NSArray *allSeries = [NSArray array];
